# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Estore API",
        "description": "Estore Api",
        "contact": {
            "name": "Victor Maamaa",
            "email": "maaamaavictor@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get a list of customers",
                "operationId": "1b873d7b0c63876f30a978b52c4d24d2",
                "parameters": [
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of items per page",
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Customers retrieved successfully",
                                    "data": {
                                        "current_page": 1,
                                        "data": [
                                            {
                                                "id": 1,
                                                "name": "John Doe",
                                                "email": "john.doe@example.com",
                                                "photo": "https://example.com/profile.jpg",
                                                "gender": "male",
                                                "phone": "+1234567890",
                                                "birthday": "1990-01-15T00:00:00.000000Z",
                                                "created_at": "2024-01-08T18:44:16.000000Z",
                                                "updated_at": "2024-01-08T18:44:16.000000Z",
                                                "deleted_at": null
                                            }
                                        ],
                                        "first_page_url": "http://127.0.0.1:8000/api/customers?page=1",
                                        "from": 1,
                                        "last_page": 1,
                                        "last_page_url": "http://127.0.0.1:8000/api/customers?page=1",
                                        "links": [
                                            {
                                                "url": null,
                                                "label": "&laquo; Previous",
                                                "active": false
                                            },
                                            {
                                                "url": "http://127.0.0.1:8000/api/customers?page=1",
                                                "label": "1",
                                                "active": true
                                            },
                                            {
                                                "url": null,
                                                "label": "Next &raquo;",
                                                "active": false
                                            }
                                        ],
                                        "next_page_url": null,
                                        "path": "http://127.0.0.1:8000/api/customers",
                                        "per_page": 10,
                                        "prev_page_url": null,
                                        "to": 1,
                                        "total": 1
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error message",
                                    "errors": []
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create a new customer",
                "operationId": "747630c757eae4c38404384330e68c57",
                "requestBody": {
                    "description": "Customer creation data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "name": "John Doe",
                                "email": "john.doe@example.com",
                                "photo": "https://example.com/profile.jpg",
                                "gender": "male",
                                "phone": "+1234567890",
                                "birthday": "1990-01-15T00:00:00.000000Z"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Customer created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Customer created successfully",
                                    "data": {
                                        "name": "John Doe",
                                        "email": "john.doe@example.com",
                                        "photo": "https://example.com/profile.jpg",
                                        "gender": "male",
                                        "phone": "+1234567890",
                                        "birthday": "1990-01-15T00:00:00.000000Z",
                                        "created_at": "2024-01-08T18:44:32.000000Z",
                                        "updated_at": "2024-01-08T18:44:32.000000Z",
                                        "id": 1
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error message",
                                    "errors": []
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/customers/{customer}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get details of a specific customer",
                "operationId": "759c7d0a3318870b151c5d02cf5e7108",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Customer name",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "email",
                        "in": "query",
                        "description": "Customer email",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "gender",
                        "in": "query",
                        "description": "Customer gender",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "male",
                                "female"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Customer retrieved successfully",
                                    "data": {
                                        "id": 1,
                                        "name": "John Doe",
                                        "email": "john.doe@example.com",
                                        "photo": "https://example.com/profile.jpg",
                                        "gender": "male",
                                        "phone": "+1234567890",
                                        "birthday": "1990-01-15T00:00:00.000000Z",
                                        "created_at": "2024-01-08T18:44:16.000000Z",
                                        "updated_at": "2024-01-08T18:44:16.000000Z",
                                        "deleted_at": null
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error message",
                                    "errors": []
                                }
                            }
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update details of a specific customer",
                "operationId": "1cdec79c0426d12f4c705878a1dbd105",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Customer update data",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            },
                            "example": {
                                "name": "Updated John Doe",
                                "email": "updated.john.doe@example.com",
                                "photo": "https://example.com/updated-profile.jpg",
                                "gender": "male",
                                "phone": "+1234567890",
                                "birthday": "1990-01-15"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Customer updated successfully",
                                    "data": {
                                        "id": 1,
                                        "name": "Updated John Doe",
                                        "email": "updated.john.doe@example.com",
                                        "photo": "https://example.com/updated-profile.jpg",
                                        "gender": "male",
                                        "phone": "+1234567890",
                                        "birthday": "1990-01-15T00:00:00.000000Z",
                                        "created_at": "2024-01-08T18:44:16.000000Z",
                                        "updated_at": "2024-01-08T18:44:16.000000Z",
                                        "deleted_at": null
                                    }
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error message",
                                    "errors": []
                                }
                            }
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Delete a specific customer",
                "operationId": "d9b736a16f6aa96eed5c485bc2346600",
                "parameters": [
                    {
                        "name": "customer",
                        "in": "path",
                        "description": "Customer ID",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": true,
                                    "message": "Customer Deleted successfully",
                                    "data": []
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "success": false,
                                    "message": "Error message",
                                    "errors": []
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "sanctum": []
        }
    ]
}